+++ modules/perception/fusion/lib/data_association/hm_data_association/chi_squared_cdf_2_0.0500_0.999900.h
@@ -82,7 +82,7 @@ inline double ChiSquaredCdf2TableFun(double dist) {
   static const size_t table_size =
       sizeof(ChiSquaredCdf2TableTable) / sizeof(double);
   double dist_ind = dist / step;
-  int dist_int = dist_ind;
+  int dist_int = static_cast<int>(dist_ind);
   double w = dist_ind - dist_int;
   if (dist_ind >= table_size - 1) {
     return 1.0;
